CORE-2057
Edit Button Save now determines if only a singular button sync message
needs to be sent - when logged in as head office and the Button Grid
already exists - the new BUTTON SyncMessageType value is used instead of
BUTTON_GRID when these conditions are met. Minor tidying carried out on
these functions to reduce indentation and avoid code duplications.

Added a new button to the Button Grid view to send a BUTTON_GRID sync
command to tills.