Combine else { if } to an else if {}
Change hardcoded use of "image/png" to use of MediaType.IMAGE_PNG constant

When erroring a button save, for any reason except file size issues, check if there was a image uploaded and set the buttonImage to that uploaded image. Then in JavaScript detect this behaviour and automatically add the image to the upload form.