CORE-3264 CORE-3287
Fixed an issue in that editing a variant on the same day was already creating a new one so you'd end up with multiple variants with identical effective dates. Now we update the record instead.
Removed some incorrect duplicate SKU validation.
Ensured the current retail and cost prices are displayed on the product search (not future effective dated changes).