STMP-100
added front end validation for max character lengths
checking that typed values are numeric and within range for store number.

groovy validation used to check for numeric on phone number due to constraint in groovy fields not allowing for phone number regex pattern to be used in front end.
same for postcode